examples/python/requirements.txt
bleak
kaitaistruct
pycryptodomex
